feat: wire calculation engine into module-calculator with flat coroutine pipeline
- Move calculation infrastructure from module-app to module-infra:
  V4 decorator chain, cube engine, PresetCalculationHelper, PureExpectationCalculator,
  StarforceLookupAdapter, CubeEngineFeatureFlag, TableMassConfig
- Add CubeTrialsPort core interface with CubeServiceImpl adapter
- Add CalculationPortConfig for shared calculation beans
  (CubeRateCalculator, StatParser, ProbabilityConvolver, FlameScoreResolver, etc.)
- Rewrite SnapshotChunkProcessor with flat coroutine pipeline:
  flatten all items across presets into Channel<FlatItem>,
  process each item independently via EquipmentExpectationCalculatorFactory
  with single Dispatchers.Default — no parallelStream/ForkJoinPool nesting
- Use @Import for selective bean loading in CalculatorApplication
- Result: 35K items in 4.4s (2.4x faster than parallelStream approach)
