--- module-app/src/main/java/maple/expectation/application/service/calculator/v4/impl/StarforceDecoratorV4.java
@@ -1,124 +1,0 @@
-package maple.expectation.application.service.calculator.v4.impl;
-
-import java.util.Optional;
-import maple.expectation.application.service.calculator.v4.EquipmentEnhanceDecorator;
-import maple.expectation.application.service.calculator.v4.EquipmentExpectationCalculator;
-import maple.expectation.core.calculator.port.StarforceLookupPort;
-
-/**
- * V4 스타포스 데코레이터 (#240)
- *
- * <h3>스타포스 특성</h3>
- *
- * <ul>
- *   <li>장비의 기본 스탯 강화 (0~25성)
- *   <li>15성 이상부터 파괴 확률 존재
- *   <li>세이프가드 옵션으로 파괴 방지 가능 (비용 2배)
- * </ul>
- *
- * <h3>Lookup Table 사용</h3>
- *
- * <p>StarforceLookupTable에서 pre-computed 기대값 조회 (O(1))
- *
- * @see StarforceLookupPort 스타포스 기대값 조회
- * @see EquipmentEnhanceDecorator 추상 데코레이터
- */
-public class StarforceDecoratorV4 extends EquipmentEnhanceDecorator {
-
-  private static final int PRECISION_SCALE = 2;
-
-  private final StarforceLookupPort lookupPort;
-  private final int currentStar;
-  private final int targetStar;
-  private final int itemLevel;
-  private Double starforceCost;
-
-  /**
-   * 스타포스 데코레이터 생성
-   *
-   * @param target 이전 단계 계산기
-   * @param lookupPort 스타포스 기대값 조회 테이블
-   * @param currentStar 현재 스타포스 (0~25)
-   * @param targetStar 목표 스타포스 (currentStar ~ 25)
-   * @param itemLevel 아이템 레벨
-   */
-  public StarforceDecoratorV4(
-      EquipmentExpectationCalculator target,
-      StarforceLookupPort lookupPort,
-      int currentStar,
-      int targetStar,
-      int itemLevel) {
-    super(target);
-    this.lookupPort = lookupPort;
-    this.currentStar = currentStar;
-    this.targetStar = targetStar;
-    this.itemLevel = itemLevel;
-  }
-
-  /** 레벨별 최대 스타로 생성 (레벨에 따라 자동 결정) */
-  public StarforceDecoratorV4(
-      EquipmentExpectationCalculator target,
-      StarforceLookupPort lookupPort,
-      int currentStar,
-      int itemLevel) {
-    this(target, lookupPort, currentStar, lookupPort.getMaxStarForLevel(itemLevel), itemLevel);
-  }
-
-  @Override
-  public double calculateCost() {
-    double previousCost = super.calculateCost();
-    double sfCost = calculateStarforceCost();
-    return previousCost + sfCost;
-  }
-
-  /**
-   * 스타포스 기대 비용 계산
-   *
-   * <p>Lookup Table에서 pre-computed 값을 조회합니다.
-   *
-   * @return 스타포스 기대 비용
-   */
-  public double calculateStarforceCost() {
-    if (starforceCost == null) {
-      // 이미 목표 스타에 도달한 경우
-      if (currentStar >= targetStar) {
-        starforceCost = 0.0;
-      } else {
-        // Lookup Table에서 기대값 조회 (이미 Double 반환)
-        starforceCost = lookupPort.getExpectedCost(currentStar, targetStar, itemLevel);
-      }
-    }
-    return starforceCost;
-  }
-
-  @Override
-  public Optional<Double> getTrials() {
-    // 스타포스는 단순 시도 횟수로 표현하기 어려움 (성공/실패/파괴 복합 확률)
-    // 대신 calculateStarforceCost()를 사용하여 기대 비용 제공
-    return Optional.empty();
-  }
-
-  @Override
-  public CostBreakdown getDetailedCosts() {
-    CostBreakdown base = super.getDetailedCosts();
-    double sfCost = calculateStarforceCost();
-    return base.withStarforce(base.starforceCost() + sfCost);
-  }
-
-  @Override
-  public String getEnhancePath() {
-    return super.getEnhancePath() + String.format(" > 스타포스(%d→%d성)", currentStar, targetStar);
-  }
-
-  public int getCurrentStar() {
-    return currentStar;
-  }
-
-  public int getTargetStar() {
-    return targetStar;
-  }
-
-  public int getItemLevel() {
-    return itemLevel;
-  }
-}
